reco_svc/requirements.txt
fastapi
uvicorn[standard]
pymysql
python-dotenv
pillow
pytesseract
httpx
pyzbar